requirements.txt
# 核心依赖
fastapi>=0.115.8
uvicorn>=0.34.0
pydantic>=2.10.6
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# 数据库相关
sqlalchemy>=2.0.30
alembic>=1.13.1
lancedb>=0.5.7

# 测试工具
pytest>=8.3.3
httpx>=0.28.1

# 数据处理
numpy>=1.26.4
pandas>=2.2.3

# 文件处理
python-multipart>=0.0.20
aiofiles>=23.2.1

# 工具库
tenacity>=9.0.0
pyyaml>=6.0.2

# AI 相关
litellm>=1.61.20

# Langchain 相关
langchain-core>=0.3.40
langchain>=0.3.19
langchain-text-splitters>=0.3.6
langchain-community>=0.3.18
langchain-experimental>=0.3.5rc1
langchain-deepseek>=0.1.2
langchain-huggingface>=0.1.2
langchain-milvus>=0.1.8
langchain-mongodb>=0.5.0
langchain-ollama>=0.2.3
langchain-openai>=0.3.7
langchain-postgres>=0.0.13
langchain-redis>=0.2.0
langchain-anthropic>=0.3.8
langchain-groq>=0.2.4 